added accessor to Respoitory Class called current_target
it defaults to a '' unless there is a single repo in the current profile in which case it defaults to the name of it.

It will check the current_target value when repos are added or removed and reset to the default value if the current value is no longer valid.

current_target is meant to store the name of the 'default' repository for the current profile in order to make addressing the specific repository easier from the command line.

Added current_target to app
  called with no argument it outputs to current_value
  called with a repo name it sets it

Specs added and all green